♻️refactor: [모의면접] 자소서 선택과 질문선택 페이지 분리
- 기존의 합쳐진 자소서 선택과 질문선택 페이지 분리
- 리스트 애니메이션 모듈로 따로 선언

KAN-85
